requirements.txt
dash>=2.0.0
openpyxl>=3.0.0
pandas>=1.3.0
pydantic>=2.0
pyyaml>=6.0
requests>=2.28.0
ruff>=0.15.5
pytest==9.0.2